refactor(tui): move provider key testing to shared module
Extract provider key testing logic from `src/tui/key-handler.js` into new shared module `src/core/provider-key-tester.js`. The TUI now imports and uses these shared functions, reducing code duplication and aligning with the Web Dashboard's `/api/key/:provider/test` endpoint implementation.

Removed from key-handler:
- `PROVIDER_TEST_MODEL_OVERRIDES`, `SETTINGS_TEST_MAX_ATTEMPTS`, `SETTINGS_TEST_RETRY_DELAY_MS`
- `PROVIDER_AUTH_ENDPOINTS` mapping
- Functions: `buildProviderModelsUrl`, `parseProviderModelIds`, `listProviderTestModels`, `classifyProviderTestOutcome`, `buildProviderTestDetail`

Updated file documentation to reflect delegation to shared module.
